requirements_torch.txt
torch>=2.2,<3.0
torchvision>=0.17,<1.0
numpy>=1.24
Pillow>=9.5
PyYAML>=6.0
pytest>=7.0
